chore(project): updated all dependencies (#38)
* chore(project): updated all dependencies

* tsconfig and type fixes

* chore(vscode): cleaned up launch.json

* build(frontend-core): added relative ci to prod bundle

* chore(frontend-core): removed duplicated relatice-ci plugin

+++ services/media/src/actions/upload-subtitles.ts
@@ -50,11 +50,13 @@ export const UploadSubtitles: RequestAction<{ urlParams: { movieId: string; acce
 
   await Promise.all(
     files
-      .filter((f) => extname(f.name) === 'vtt')
+      .filter((f) => !(f instanceof Array) && extname(f.name) === 'vtt')
       .map(async (file) => {
-        await promises.copyFile(file.path, join(targetPath, file.name))
-        // Remove from temp
-        promises.unlink(file.path)
+        if (!(file instanceof Array)) {
+          await promises.copyFile(file.path, join(targetPath, file.name))
+          // Remove from temp
+          promises.unlink(file.path)
+        }
       }),
   )
